+++ Sources/KeyboardKit/_Deprecated/KeyboardType+Deprecated.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+public extension KeyboardType {
+    
+    @available(*, deprecated, message: "Just set keyboardType directly.")
+    func canBeReplaced(with type: KeyboardType) -> Bool {
+        guard
+            case .alphabetic(let state) = self,
+            case .alphabetic(let newState) = type
+        else { return true }
+        if state == .capsLocked && newState == .uppercased { return false }
+        return true
+    }
+}