+++ Tests/KeyboardKitTests/Keyboard/KeyboardTypeTests.swift
@@ -33,25 +33,5 @@ class KeyboardTypeTests: QuickSpec {
                 expect(result(for: .custom(""))).to(beFalse())
             }
         }
-
-        
-        describe("can be replaced with keyboard type") {
-            
-            func result(from type: KeyboardType, to newType: KeyboardType) -> Bool {
-                type.canBeReplaced(with: newType)
-            }
-            
-            it("can change for most combinations") {
-                expect(result(from: .alphabetic(.lowercased), to: .alphabetic(.uppercased))).to(beTrue())
-                expect(result(from: .alphabetic(.uppercased), to: .alphabetic(.lowercased))).to(beTrue())
-                expect(result(from: .alphabetic(.capsLocked), to: .alphabetic(.lowercased))).to(beTrue())
-                expect(result(from: .emojis, to: .numeric)).to(beTrue())
-            }
-            
-            it("cannot change from alpha caps to alpha upper") {
-                expect(result(from: .alphabetic(.capsLocked), to: .alphabetic(.uppercased))).to(beFalse())
-            }
-        }
-
     }
 }